+++ playbooks/env_variables.py
@@ -1,0 +1,6 @@
+
+login_user = 'madeasy_superuser'
+login_password = 'ready+player-one'
+db_user = 'madeasy_user'
+db_pass = 'madeasy'
+db_name = 'madeasy'